updated Fetching
